Add support for self-service (#6)
This PR adds support for "self-service" visitor passes. This is a specific use case where you want to share a URL, but you don't know the contact details of the user with whom you want to share it in advance. The standard model for visitor passes is that you generate a pass for a user and then send them a unique link. If you want something that is more like a public URL, but you also want some control over its use, self-service is an option. In this use case the view is decorated, but instead of bumping the user with a 403, they are redirected to a form that asks them for their contact details, sends them a visitor pass link with the appropriate scope.