Refactor views to make them easier to override (#8)
The current implementation has a single fixed template for the visitor
request / success pages, and real-world use has shown that this may not
be flexible enough. This PR updates the way that the views work to
enable subclasses to have full control over the templates / contexts
that they use. This means you can have different templates for different
use cases.